requirements.txt
streamlit
openai
pandas
python-docx